[RNKC-062] - move codegen definitions into separate folder
