+++ src/specs/NativeStatusBarManagerCompat.ts
